重构时间变量声明和矩阵操作函数
在 `cuda_matrix.cu` 中移除多个 `extern clock_t` 变量声明，并在 `cuda_matrix.cuh` 中添加相应声明，使用条件编译避免重复定义。更新 `kernel_function.cu` 和 `kernel_function.cuh` 中 `elementwise_equal_kernel` 函数的参数类型，从 `uint32_t` 改为 `size_t`，以支持更大的数据集。

+++ StaticCudaLib/kernel_function.cu
@@ -119,7 +119,7 @@ template<typename Type> __global__ static void matrix_transpose_kernel
 }
 
 template <typename Type> __global__ void elementwise_equal_kernel
-(const Type* src1, const Type* src2, char* res, const uint32_t size)
+(const Type* src1, const Type* src2, char* res, const size_t size)
 {
 	int idx = blockIdx.x * blockDim.x + threadIdx.x;
 	if (idx < size)
